Modify ELK handle commands
