+++ src/modes/user/simple_behavior.hpp
@@ -87,7 +87,7 @@ void button_hold_default(const uint8_t clicks, const bool isEndOfHoldEvent, cons
   }
 }
 
-void bluetooth_switch_pattern(uint8_t index, uint32_t targetColor)
+void handle_elk_command(const utils::ELK::Package&)
 {
   //
   return;
