+++ src/user/cct_functions.cpp
@@ -171,7 +171,7 @@ bool should_spawn_thread() { return false; }
 
 void user_thread() {}
 
-void bluetooth_switch_pattern(uint8_t, uint32_t) {}
+void handle_elk_command(const utils::ELK::Package&) {}
 
 } // namespace lampda::user
